refactor(ir): move Fp8Format beside the fp8 scalar types
It describes the formats, so the cast lowering, the CUDA and the SPIR-V
converters all import it from the types, not from a post-processing pass.
